Added role support to all methods in resources, deleted some useless methods

+++ src/main/java/com/frizzer/employeeapp/controller/resource/EmployeeInfoResource.java
@@ -1,8 +1,11 @@
 package com.frizzer.employeeapp.controller.resource;
 
 import com.frizzer.employeeapp.entity.EmployeeInfoDto;
+import com.frizzer.employeeapp.entity.EmployeeRole;
 import com.frizzer.employeeapp.service.EmployeeInfoService;
 import com.frizzer.employeeapp.service.mapper.EmployeeInfoMapper;
+import jakarta.annotation.security.DeclareRoles;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.ejb.EJB;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
@@ -12,32 +15,48 @@ import jakarta.ws.rs.PUT;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.Response.Status;
+import jakarta.ws.rs.core.SecurityContext;
 
 @Path("/employees-info")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
+@DeclareRoles({"ADMIN", "WORKER"})
 public class EmployeeInfoResource {
 
   @EJB
   private EmployeeInfoService employeeInfoService;
 
+  @Context
+  private SecurityContext context;
+
   @POST
   @Path("/{id}")
+  @RolesAllowed("ADMIN")
   public Response save(@PathParam("id") Long id, EmployeeInfoDto employee) {
-    return Response
-        .ok(employeeInfoService.save(EmployeeInfoMapper.INSTANCE.fromDto(employee), id))
-        .build();
+    if (isAdmin()) {
+      return Response
+          .ok(employeeInfoService.save(EmployeeInfoMapper.INSTANCE.fromDto(employee), id))
+          .build();
+    } else {
+      return Response.status(Status.FORBIDDEN).build();
+    }
   }
 
   @PUT
   @Path("/{id}")
+  @RolesAllowed("ADMIN")
   public Response update(@PathParam("id") Long id, EmployeeInfoDto employee) {
-    return Response
-        .ok(employeeInfoService.update(EmployeeInfoMapper.INSTANCE.fromDto(employee), id))
-        .build();
+    if (isAdmin()) {
+      return Response
+          .ok(employeeInfoService.update(EmployeeInfoMapper.INSTANCE.fromDto(employee), id))
+          .build();
+    } else {
+      return Response.status(Status.FORBIDDEN).build();
+    }
   }
 
   @GET
@@ -50,16 +69,19 @@ public class EmployeeInfoResource {
 
   @DELETE
   @Path("/{id}")
+  @RolesAllowed("ADMIN")
   public Response delete(@PathParam("id") Long id) {
-    boolean entityExists = employeeInfoService.delete(id);
-    return entityExists ? Response.ok(Status.OK).build()
-        : Response.status(Status.NOT_FOUND).build();
+    if (isAdmin()) {
+      boolean entityExists = employeeInfoService.delete(id);
+      return entityExists ? Response.ok(Status.OK).build()
+          : Response.status(Status.NOT_FOUND).build();
+    } else {
+      return Response.status(Status.FORBIDDEN).build();
+    }
   }
 
-  @GET
-  @Path("/hello")
-  public String hello() {
-    return "Hello,i'm currently live";
+  private boolean isAdmin() {
+    return context.isUserInRole(String.valueOf(EmployeeRole.ADMIN));
   }
 
 }
